+++ lib/src/utils/youtube_player_flags.dart
@@ -1,4 +1,4 @@
-// Copyright 2019 Sarbagya Dhaubanjar. All rights reserved.
+// Copyright 2020 Sarbagya Dhaubanjar. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -64,6 +64,11 @@ class YoutubePlayerFlags {
   /// Default is `en`.
   final String captionLanguage;
 
+  /// Forces High Definition video quality when possible
+  ///
+  /// Default is false.
+  final bool forceHD;
+
   /// Creates [YoutubePlayerFlags].
   const YoutubePlayerFlags({
     this.hideControls = false,
@@ -77,6 +82,7 @@ class YoutubePlayerFlags {
     this.enableCaption = true,
     this.captionLanguage = 'en',
     this.loop = false,
+    this.forceHD = false,
   });
 
   /// Copies new values assigned to the [YoutubePlayerFlags].
@@ -91,6 +97,7 @@ class YoutubePlayerFlags {
     bool disableDragSeek,
     bool loop,
     bool enableCaption,
+    bool forceHD,
     String captionLanguage,
   }) {
     return YoutubePlayerFlags(
@@ -104,6 +111,7 @@ class YoutubePlayerFlags {
       isLive: isLive ?? this.isLive,
       loop: loop ?? this.loop,
       mute: mute ?? this.mute,
+      forceHD: forceHD ?? this.forceHD,
     );
   }
 }