+++ lib/src/widgets/play_pause_button.dart
@@ -69,18 +69,11 @@ class _PlayPauseButtonState extends State<PlayPauseButton>
 
   @override
   Widget build(BuildContext context) {
-    if (_controller.value.playerState == PlayerState.buffering) {
-      return widget.bufferIndicator ??
-          Container(
-            width: 70.0,
-            height: 70.0,
-            child: CircularProgressIndicator(
-              valueColor: AlwaysStoppedAnimation(Colors.white),
-            ),
-          );
-    } else {
+    final _playerState = _controller.value.playerState;
+    if (_playerState == PlayerState.playing ||
+        _playerState == PlayerState.paused) {
       return Visibility(
-        visible: _controller.value.playerState == PlayerState.cued ||
+        visible: _playerState == PlayerState.cued ||
             !_controller.value.isPlaying ||
             _controller.value.isControlsVisible,
         child: Material(
@@ -100,5 +93,13 @@ class _PlayPauseButtonState extends State<PlayPauseButton>
         ),
       );
     }
+    return widget.bufferIndicator ??
+        Container(
+          width: 70.0,
+          height: 70.0,
+          child: CircularProgressIndicator(
+            valueColor: AlwaysStoppedAnimation(Colors.white),
+          ),
+        );
   }
 }